Move a few more GitSCM tests
